+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Fix to render recursively in worker
 
 ## [0.0.1] - 2020-12-28
 ### Added